src/requirements.txt
allure-behave==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
behave==1.2.6
greenlet==1.1.2
parse==1.19.0
parse-type==0.6.0
playwright==1.19.1
pluggy==1.0.0
pyee==8.1.0
PyYAML==6.0
six==1.16.0
websockets==10.1
